requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
onnxruntime>=1.20.0
opencv-python==4.8.0.76
pillow==10.1.0
numpy>=1.26.0
requests==2.31.0
python-multipart==0.0.8
ultralytics>=8.0.0
